Clean up integer types for porting to 32-bit

+++ dvd_cell.c
@@ -4,7 +4,7 @@
  * Functions used to get information about a DVD cell
  */
 
-uint32_t dvd_cell_first_sector(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number) {
+uint64_t dvd_cell_first_sector(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number) {
 
 	if(vts_ifo->vts_pgcit == NULL || vts_ifo->vts_ptt_srpt == NULL || vts_ifo->vts_ptt_srpt->title == NULL)
 		return 0;
@@ -16,14 +16,14 @@ uint32_t dvd_cell_first_sector(const ifo_handle_t *vmg_ifo, const ifo_handle_t *
 	if(pgc == NULL || pgc->program_map == NULL)
 		return 0;
 	
-	uint32_t sector = 0;
-	sector = pgc->cell_playback[cell_number - 1].first_sector;
+	uint64_t sector = 0;
+	sector = (uint64_t)pgc->cell_playback[cell_number - 1].first_sector;
 
 	return sector;
 
 }
 
-uint32_t dvd_cell_last_sector(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number) {
+uint64_t dvd_cell_last_sector(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number) {
 
 	if(vts_ifo->vts_pgcit == NULL || vts_ifo->vts_ptt_srpt == NULL || vts_ifo->vts_ptt_srpt->title == NULL)
 		return 0;
@@ -35,23 +35,23 @@ uint32_t dvd_cell_last_sector(const ifo_handle_t *vmg_ifo, const ifo_handle_t *v
 	if(pgc == NULL || pgc->program_map == NULL)
 		return 0;
 	
-	uint32_t sector = 0;
-	sector = pgc->cell_playback[cell_number - 1].last_sector;
+	uint64_t sector = 0;
+	sector = (uint64_t)pgc->cell_playback[cell_number - 1].last_sector;
 
 	return sector;
 
 }
 
-ssize_t dvd_cell_blocks(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number) {
+uint64_t dvd_cell_blocks(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number) {
 
-	uint32_t first_sector = 0;
+	uint64_t first_sector = 0;
 	first_sector = dvd_cell_first_sector(vmg_ifo, vts_ifo, track_number, cell_number);
 
-	uint32_t last_sector = 0;
+	uint64_t last_sector = 0;
 	last_sector = dvd_cell_last_sector(vmg_ifo, vts_ifo, track_number, cell_number);
 
-	ssize_t blocks = 0;
-	blocks = (ssize_t)(last_sector - first_sector);
+	uint64_t blocks = 0;
+	blocks = last_sector - first_sector;
 
 	// Include the last cell
 	if(last_sector == first_sector)
@@ -61,12 +61,12 @@ ssize_t dvd_cell_blocks(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo
 
 }
 
-ssize_t dvd_cell_filesize(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number) {
+uint64_t dvd_cell_filesize(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number) {
 
-	ssize_t blocks;
+	uint64_t blocks;
 	blocks = dvd_cell_blocks(vmg_ifo, vts_ifo, track_number, cell_number);
 
-	ssize_t filesize = 0;
+	uint64_t filesize = 0;
 	filesize = blocks * DVD_VIDEO_LB_LEN;
 
 	return filesize;
@@ -81,8 +81,8 @@ bool dvd_track_min_sector_error(const ifo_handle_t *vmg_ifo, const ifo_handle_t 
 		return false;
 
 	uint8_t cell = 2;
-	uint32_t first_sector = 0;
-	uint32_t min_sector = 0;
+	uint64_t first_sector = 0;
+	uint64_t min_sector = 0;
 
 	min_sector = dvd_cell_first_sector(vmg_ifo, vts_ifo, track_number, 1);
 
@@ -109,8 +109,8 @@ bool dvd_track_max_sector_error(const ifo_handle_t *vmg_ifo, const ifo_handle_t 
 		return false;
 
 	uint8_t cell = 2;
-	uint32_t last_sector = 0;
-	uint32_t max_sector = 0;
+	uint64_t last_sector = 0;
+	uint64_t max_sector = 0;
 
 	max_sector = dvd_cell_first_sector(vmg_ifo, vts_ifo, track_number, 1);
 
@@ -137,8 +137,8 @@ bool dvd_track_repeat_first_sector_error(const ifo_handle_t *vmg_ifo, const ifo_
 		return false;
 
 	uint8_t cell = 2;
-	uint32_t first_cell_first_sector = 0;
-	uint32_t first_sector = 0;
+	uint64_t first_cell_first_sector = 0;
+	uint64_t first_sector = 0;
 
 	first_cell_first_sector = dvd_cell_first_sector(vmg_ifo, vts_ifo, track_number, 1);
 
@@ -163,8 +163,8 @@ bool dvd_track_repeat_last_sector_error(const ifo_handle_t *vmg_ifo, const ifo_h
 		return false;
 
 	uint8_t cell = 2;
-	uint32_t first_cell_last_sector = 0;
-	uint32_t last_sector = 0;
+	uint64_t first_cell_last_sector = 0;
+	uint64_t last_sector = 0;
 
 	first_cell_last_sector = dvd_cell_last_sector(vmg_ifo, vts_ifo, track_number, 1);
