+++ dvd_vob.h
@@ -7,12 +7,12 @@
 struct dvd_vob {
 	uint16_t vts;
 	uint16_t vob;
-	ssize_t blocks;
-	ssize_t filesize;
+	uint64_t blocks;
+	uint64_t filesize;
 };
 
-ssize_t dvd_vob_blocks(dvd_reader_t *dvdread_dvd, const uint16_t vts_number, const uint16_t vob_number);
+uint64_t dvd_vob_blocks(dvd_reader_t *dvdread_dvd, const uint16_t vts_number, const uint16_t vob_number);
 
-ssize_t dvd_vob_filesize(dvd_reader_t *dvdread_dvd, const uint16_t vts_number, const uint16_t vob_number);
+uint64_t dvd_vob_filesize(dvd_reader_t *dvdread_dvd, const uint16_t vts_number, const uint16_t vob_number);
 
 #endif
